Fix Area Server Events
Back in the day we did the minimal amount of work needed to get the Area Servers online, and never went back to investigate the packets. As a result the Events never worked properly.  Turns out we were missing one packet with the Lobby Server DateTime. Events should work as intended.